requirements.txt
black==22.1.0
Django==4.0.2
djangorestframework==3.13.1
pytest==7.0.1
pytest-django==4.5.2
pytz==2021.3
requests==2.27.1
sqlparse==0.4.2
Markdown==3.3.6
django-filter==21.1